@@ -4,11 +4,11 @@
   </div>
 </template>
 <script>
-import Vue from "vue";
+import { defineComponent } from "vue";
 // If you don;t have this file run `yarn` cmd to create it
 import Playground from "./Playground.vue";
-export default Vue.extend({
-  name: "App",
+export default defineComponent({
+  // name: "App",
   components: {
     Playground,
   },
