@@ -1,6 +1,5 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App.vue";
-Vue.config.productionTip = false;
-new Vue({
-  render: (h) => h(App),
-}).$mount("#app");
+
+const app = createApp(App);
+app.mount("#app");
